removed unused functions and modules ignore erroneous optimization updates
