+++ vlgp/api.py
@@ -1,24 +1,26 @@
-from numpy import asarray, newaxis
+from numpy import asarray, newaxis, empty
 
-from vlgp.experimental import initialize, vem, postprocess
+from .callback import Saver, Progressor
+from .core import initialize, vem, check_y_type
+from .util import add_constant, lagmat
 
 
 def fit(y,
-        types,
-        z_ndim,
-        x=None,
+        obs_types,
+        dyn_ndim,
+        exog=None,
         a=None,
         b=None,
-        history_filter=0,
         mu=None,
+        history_filter=0,
         z=None,
         alpha=None,
         beta=None,
         sigma=None,
         omega=None,
         rank=None,
-        eps=1e-8,
-        tol=1e-5,
+        path=None,
+        callbacks=None,
         **kwargs):
     """
     vLGP main function
@@ -27,12 +29,12 @@ def fit(y,
     ----------
     y : ndarray
         obserbation
-    types : ndarray
+    obs_types : ndarray
         types of observation dimensions, 'spike' or 'lfp'
-    z_ndim : int
+    dyn_ndim : int
         number of latent dimensions
-    x : ndarray
-        regression variables
+    exog : ndarray, optional
+        external factors
     a : ndarray, optional
         initial value of loading
     b : ndarray, optional
@@ -45,17 +47,18 @@ def fit(y,
         true value of loading
     beta : ndarray, optional
         true value of regression
-    sigma : ndarray, optional
-        initial value of prior variance
-    omega : ndarray, optional
-        initial value of prior timescale
     history_filter : int, optional
+        history_filter length
     rank : int, optional
         rank of incomplete Cholesky
     eps : double, optional
         a small positive number
     tol : double, optional
         numerical tolerance
+    path : string, optional
+        path to the save file
+    callbacks : list, optional
+        callbacks
     kwargs : dict, optional
         algorithm options. See fill_options()
 
@@ -65,36 +68,56 @@ def fit(y,
         fit
     """
 
-    model = dict()
     y = asarray(y)
     y = y.astype(float)
     if y.ndim < 2:
         y = y[..., newaxis]
     if y.ndim < 3:
         y = y[newaxis, ...]
-    model['y'] = y
-    model['z_ndim'] = z_ndim
-    model['history_filter'] = history_filter
-    model['types'] = types
-    model['x'] = x
-    model['a'] = a
-    model['b'] = b
-    model['mu'] = mu
-    model['sigma'] = sigma
-    model['omega'] = omega
-    model['rank'] = rank
-    model['z'] = z
-    model['alpha'] = alpha
-    model['beta'] = beta
 
-    # model['x'] = make_regression(x, y, history_filter)
+    obs_types = check_y_type(obs_types)
 
-    options = dict(kwargs)
-    options['eps'] = eps
-    options['tol'] = tol
+    ntrial, nbin, obs_ndim = y.shape
+
+    # make design matrix of regression
+    h = empty((obs_ndim, ntrial, nbin, 1 + history_filter), dtype=float)
+    for obs_dim in range(obs_ndim):
+        for trial in range(ntrial):
+            h[obs_dim, trial, :] = add_constant(lagmat(y[trial, :, obs_dim], lag=history_filter))
+
+    # Initialize posterior and loading
+    # Use factor analysis if both missing initial values
+    # Use least squares if missing one of loading and latent
+
+    model = dict(y=y,
+                 channel=obs_types,
+                 dyn_ndim=dyn_ndim,
+                 history_filter=history_filter,
+                 h=h,
+                 mu=mu,
+                 a=a,
+                 b=b,
+                 sigma=sigma,
+                 omega=omega,
+                 rank=rank,
+                 x=z,
+                 alpha=alpha,
+                 beta=beta,
+                 path=path,
+                 options=kwargs)
 
     initialize(model)
-    vem(model)
-    postprocess(model)
+
+    saver = Saver()
+    pbar = Progressor(model['options']['niter'])
+
+    callbacks = callbacks or []
+    callbacks.extend([pbar.update, saver.save])
+    try:
+        vem(model, callbacks)
+    finally:
+        print('Exiting...')
+        pbar.print(model)
+        saver.save(model, force=True)
 
     return model
